% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microbiome.R
\name{adonis3}
\alias{adonis3}
\title{Permutational Multivariate Analysis of Variance Using Distance Matrices}
\usage{
adonis3(
  formula,
  data,
  permutations = 999,
  method = "bray",
  sqrt.dist = FALSE,
  add = FALSE,
  by = "terms",
  parallel = getOption("mc.cores"),
  na.action = stats::na.fail,
  strata = NULL,
  ...
)
}
\arguments{
\item{formula}{Model formula.}

\item{data}{Data frame.}

\item{permutations}{Number of permutations (default 999).}

\item{method}{Distance method.}

\item{sqrt.dist}{Logical, whether to take the square root of distances.}

\item{add}{Logical or character, whether to add a constant to distances.}

\item{by}{Character, margin or terms.}

\item{parallel}{Number of parallel cores.}

\item{na.action}{Function for handling NAs.}

\item{strata}{Permutation strata.}

\item{...}{Additional arguments.}
}
\value{
An ANOVA table.
}
\description{
Permutational Multivariate Analysis of Variance Using Distance Matrices
}
