% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microbiome.R
\name{alpha_diversity}
\alias{alpha_diversity}
\title{Calculate alpha diversity indices}
\usage{
alpha_diversity(x, tree = NULL)
}
\arguments{
\item{x}{Species abundance matrix.}

\item{tree}{Optional phylogenetic tree for PD calculation.}
}
\value{
A data frame containing multiple alpha diversity metrics.
}
\description{
Calculate alpha diversity indices
}
