% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_models.R
\name{calculate_metrics}
\alias{calculate_metrics}
\title{Calculate comprehensive model metrics}
\usage{
calculate_metrics(truth, response)
}
\arguments{
\item{truth}{Numeric vector of true values.}

\item{response}{Numeric vector of predicted values.}
}
\value{
A matrix with MAE, MSE, RMSE, Rho, and R2.
}
\description{
Calculate comprehensive model metrics
}
