% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{chdir}
\alias{chdir}
\title{Change workdir more fast}
\usage{
chdir(outdir)
}
\arguments{
\item{outdir}{where workdir you wang change to.}
}
\value{
None
}
\description{
Change workdir more fast
}
\examples{
getwd()
outdir = "../"
chdir(outdir)
getwd()
}
