% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col_ttest.R
\name{col_ttest}
\alias{col_ttest}
\title{Title}
\usage{
col_ttest(data, g1, g2, adjust = TRUE, p = TRUE)
}
\arguments{
\item{data}{the data.frame colume has the two group need to test by t.test}

\item{g1}{the number range for group1 like "`1:3`"}

\item{g2}{the number range for group2 like "`4:6`"}

\item{adjust}{TRUE/FALSE add/ or not add the qvalue by qvalue packsages.}

\item{p}{TRUE/FALSE add/ or not add the pvalue by ttest.}
}
\value{
the qvalue and pvalue
}
\description{
Title
}
\examples{
test = matrix(rnorm(200), 20, 10)
test[1:10, seq(1, 10, 2)] = test[1:10, seq(1, 10, 2)] + 3
test[11:20, seq(2, 10, 2)] = test[11:20, seq(2, 10, 2)] + 2
test[15:20, seq(2, 10, 2)] = test[15:20, seq(2, 10, 2)] + 4
colnames(test) = paste("Test", 1:10, sep = "")
rownames(test) = paste("Gene", 1:20, sep = "")
col_ttest(data = test,g1 = 1:5,g2 = 6:10)
}
