% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_models.R
\name{cor_spearman}
\alias{cor_spearman}
\title{Calculate correlation between microbiome and metabolites}
\usage{
cor_spearman(microbiome, metabolites, name, method = "spearman")
}
\arguments{
\item{microbiome}{Dataframe of microbiome abundance.}

\item{metabolites}{Dataframe of metabolites abundance.}

\item{name}{Character vector of length 2 for column names.}

\item{method}{Character, correlation method ("spearman", "pearson").}
}
\value{
A list containing association data, correlations, q-values, and p-values.
}
\description{
Calculate correlation between microbiome and metabolites
}
