% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_models.R
\name{diffslope}
\alias{diffslope}
\title{Calculate difference in slopes between two linear models}
\usage{
diffslope(
  x1,
  y1,
  x2,
  y2,
  permutations = 1000,
  ic = FALSE,
  resc.x = FALSE,
  resc.y = TRUE,
  trace = FALSE,
  ...
)
}
\arguments{
\item{x1}{Numeric vector for x of group 1.}

\item{y1}{Numeric vector for y of group 1.}

\item{x2}{Numeric vector for x of group 2.}

\item{y2}{Numeric vector for y of group 2.}

\item{permutations}{Integer, number of permutations.}

\item{ic}{Logical, whether to calculate intercept difference.}

\item{resc.x}{Logical, rescale x.}

\item{resc.y}{Logical, rescale y.}

\item{trace}{Logical, print trace.}

\item{...}{Additional arguments.}
}
\value{
An S3 object of class `dsl` or `dsl2`.
}
\description{
Calculate difference in slopes between two linear models
}
