% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_df.R
\name{get_df}
\alias{get_df}
\title{Title}
\usage{
get_df(data, var = NULL, head = FALSE)
}
\arguments{
\item{data}{tibble need to become a data.frame objsct}

\item{var}{which colume will be rownames,the default is the first column}

\item{head}{if TRUE, will display the data like head()}
}
\value{
a data.frame
}
\description{
Title
}
\examples{
print("creat a example to test it")
}
