% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mat.R
\name{get_mat}
\alias{get_mat}
\title{Title the function data 2 matrix}
\usage{
get_mat(df, ...)
}
\arguments{
\item{df}{the the input must be a data.frame, tibble or matrix}

\item{...}{inhert parameter}
}
\value{
a matrix
}
\description{
Title the function data 2 matrix
}
\examples{
test = matrix(rnorm(200), 20, 10)
colnames(test) = paste("Test", 1:10, sep = "")
rownames(test) = paste("Gene", 1:20, sep = "")
df = as.data.frame(test)
str(df)
mat = get_mat(df)
str(mat)
}
