% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tb.R
\name{get_tb}
\alias{get_tb}
\title{Title}
\usage{
get_tb(data, var = "id", head = FALSE)
}
\arguments{
\item{data}{data.frame need to become a tibble objsct}

\item{var}{the rownames col's names, the default is "id"}

\item{head}{if TRUE, will display the data like head()}
}
\value{
tibble object
}
\description{
Title
}
\examples{
print("creat a example to test it")
}
