% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/koid_to_pathway.R
\name{koid_to_pathway}
\alias{koid_to_pathway}
\title{Title}
\usage{
koid_to_pathway(koid)
}
\arguments{
\item{koid}{the list or vctor of ko id like "ko00010"}
}
\value{
a data.frame with three column:"KEEG_A","KEGG_B","desc"
}
\description{
Title
}
\examples{
#koid_to_pathway("ko00010")
}
