% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.r
\name{messageline}
\alias{messageline}
\title{Title nothing}
\usage{
messageline(message)
}
\arguments{
\item{message}{you want say to user}
}
\value{
character
}
\description{
Title nothing
}
\examples{
messageline("yesimola !")
}
