% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_models.R
\name{metric_mse}
\alias{metric_mse}
\title{Calculate Mean Squared Error (MSE)}
\usage{
metric_mse(actual, predicted)
}
\arguments{
\item{actual}{Numeric vector of true values.}

\item{predicted}{Numeric vector of predicted values.}
}
\description{
Calculate Mean Squared Error (MSE)
}
