% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_models.R
\name{metric_rmse}
\alias{metric_rmse}
\title{Calculate Root Mean Squared Error (RMSE)}
\usage{
metric_rmse(actual, predicted)
}
\arguments{
\item{actual}{Numeric vector of true values.}

\item{predicted}{Numeric vector of predicted values.}
}
\description{
Calculate Root Mean Squared Error (RMSE)
}
