% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_models.R
\name{metric_rsquared}
\alias{metric_rsquared}
\title{Calculate R-squared}
\usage{
metric_rsquared(actual, predicted)
}
\arguments{
\item{actual}{Numeric vector of true values.}

\item{predicted}{Numeric vector of predicted values.}
}
\description{
Calculate R-squared
}
