% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microbiome.R
\name{read_mib}
\alias{read_mib}
\title{Read and align microbiome abundance and metadata}
\usage{
read_mib(file_path, verbose = FALSE)
}
\arguments{
\item{file_path}{Character, path to the excel file.}

\item{verbose}{Logical, print matching report or not.}
}
\value{
A list containing aligned `exp` and `meta` dataframes.
}
\description{
Read and align microbiome abundance and metadata
}
