% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_dat.R
\name{t_dat}
\alias{t_dat}
\title{Title}
\usage{
t_dat(data)
}
\arguments{
\item{data}{a data.frame need to transpose the columns and rows}
}
\value{
data.frame with the pvalue and qvalue
}
\description{
Title
}
\examples{
test = matrix(rnorm(200), 20, 10)
test[1:10, seq(1, 10, 2)] = test[1:10, seq(1, 10, 2)] + 3
test[11:20, seq(2, 10, 2)] = test[11:20, seq(2, 10, 2)] + 2
test[15:20, seq(2, 10, 2)] = test[15:20, seq(2, 10, 2)] + 4
colnames(test) = paste("Test", 1:10, sep = "")
rownames(test) = paste("Gene", 1:20, sep = "")
t_dat(test)
}
