% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method.R
\name{xplot.heatpoint}
\alias{xplot.heatpoint}
\title{Title}
\usage{
\method{xplot}{heatpoint}(obj, ...)
}
\arguments{
\item{obj}{the input data for the function `heatpoint`.}

\item{...}{Other parameters}
}
\value{
a ggplot2 object
}
\description{
Title
}
\examples{
\dontrun{
points = 5000
x = c(rnorm(points/2))
y = x + rnorm(points/2,sd=0.8)
x = sign(x)*abs(x)^1.3
heatpoint(x,y) -> dat_result
xplot(dat_result)
}
}
