% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method.R
\name{xplot.rangecount}
\alias{xplot.rangecount}
\title{Title the plot fucntion for rangecount}
\usage{
\method{xplot}{rangecount}(obj, ...)
}
\arguments{
\item{obj}{object}

\item{...}{Other parameters}
}
\value{
ggplot object
}
\description{
Title the plot fucntion for rangecount
}
\examples{
# nothing
}
