% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method.R
\name{xplot}
\alias{xplot}
\title{Title add the new function for plot the result from other function.}
\usage{
xplot(obj, ...)
}
\arguments{
\item{obj}{object}

\item{...}{other parameter}
}
\value{
s3 methods
}
\description{
Title add the new function for plot the result from other function.
}
\examples{
# nothing
}
