% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/need_color.R
\name{disco}
\alias{disco}
\title{Title display the color which you choose with the function 'need_color'}
\usage{
disco(colpal = NULL, ncol = NULL)
}
\arguments{
\item{colpal}{a valid color palette name(color number + id,such as 'two1',can see disco())
or a numberic vector that can select color randomly.}

\item{ncol}{when set the colpal names, ncol set how many color you need}
}
\value{
plot a plot display the color you choose
}
\description{
Title display the color which you choose with the function 'need_color'
}
\examples{
# nothing

}
