% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method.R
\name{print.rangecount}
\alias{print.rangecount}
\title{Title change the print method for range_count function result.}
\usage{
\method{print}{rangecount}(obj, ...)
}
\arguments{
\item{obj}{objsct}

\item{...}{Other parameters}
}
\value{
a s3 object
}
\description{
Title change the print method for range_count function result.
}
\examples{
# nothing
}
